requirements.txt
Flask
Flask-CORS
python-dotenv
mysql-connector-python
Werkzeug
